requirements.txt
Django~=5.0.7
django-rest-authemail~=2.1.7